+++ include/core/constants.h
@@ -1,0 +1,5 @@
+#ifndef CONSTANTS_H
+#define CONSTANTS_H
+#define PI 3.141592653589793
+
+#endif
