+++ src/core/freq_filters.cpp
